# gscores-api/src/main/java/asia/goldenowl/gscores_api/dto/ScoreDto.java
package asia.goldenowl.gscores_api.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class ScoreDto {

    private String subjectName;

    private String subjectCode;

    private Float score;

}